storage: init loop device for linbit piraeus operator

+++ scripts/delete.sh
@@ -2,17 +2,30 @@
 
 NS=${NS:-`basename "$PWD"`}
 DEPLOY=${DEPLOY:-`basename "$PWD"`}
+HELM_EXTRA=${HELM_EXTRA:-}
+DEBUG=${DEBUG:-false}
 
 echo namespace: ${NS}
 echo deployment: ${DEPLOY}
 
 if [ ! -d charts ]; then
    echo "Building depencency..."
-   helm dependency build 
+   ${DBG_COMMAND} helm dependency build
 fi
 
-helm template \
-    --namespace ${NS} \
-    ${DEPLOY} . \
-    | kubectl -n ${NS} delete -f -
+if [ -f values-extra.yaml ]; then
+    HELM_EXTRA="-f values-extra.yaml"
+fi
 
+if [ "$DEBUG" = "true" ]; then
+    echo helm template \
+        --namespace ${NS} \
+        ${HELM_EXTRA} \
+        ${DEPLOY} .
+else
+    helm template \
+        --namespace ${NS} \
+        ${HELM_EXTRA} \
+        ${DEPLOY} . \
+        | kubectl -n ${NS} delete -f -
+fi
